requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
python-dotenv==1.0.1
APScheduler==3.10.4
email-validator==2.0.0.post2